+++ src/components/Button/tests/Button.test.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { mount } from 'enzyme';
 import 'jest-styled-components';
 import Button from '../index';
-// import StyledButton from '../../../styled/Button';
 
 let renderedComponent;
 const renderComponent = (props = {}, children = null) =>
@@ -13,14 +12,6 @@ describe('<Button />', () => {
     renderedComponent.unmount();
   });
 
-  // it('should have a default type equals to button', () => {
-  //   renderedComponent = renderComponent();
-  //   const defaultType = 'button';
-  //   const wrapper = renderedComponent.find(Button);
-  //   expect(wrapper.find(StyledButton).exists()).toBe(true);
-  //   expect(renderedComponent.at(0).prop('type')).toBe(defaultType);
-  // });
-
   it('should render a child', () => {
     const Child = () => <span>child</span>;
     renderedComponent = renderComponent({}, <Child />);
@@ -28,10 +19,4 @@ describe('<Button />', () => {
     expect(wrapper.exists()).toBe(true);
     expect(wrapper.find(Child).exists()).toBe(true);
   });
-
-  // it('should display a label', () => {
-  //   renderedComponent = renderComponent({ label: 'Label' });
-  //   const wrapper = renderedComponent.find(Button);
-  //   expect(wrapper.text()).toBe('Label');
-  // });
 });
