+++ src/styled/Checkbox/index.js
@@ -1,0 +1,8 @@
+/**
+ *
+ * Checkbox
+ *
+ */
+
+export { default } from './Checkbox';
+export { default as CheckboxWrapper } from './CheckboxWrapper';
