+++ src/styled/Checkbox/tests/Checkbox.test.js
@@ -1,0 +1,11 @@
+// import React from 'react';
+// import { mount } from 'enzyme';
+// import { enzymeFind } from 'styled-components/test-utils';
+
+// import Checkbox from '../index';
+
+describe('<Checkbox />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(true);
+  });
+});
