--- src/styled/GlobalStyle/GlobalStyles.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import General from './General';
-import Font from './Font';
-
-function GlobalStyles() {
-  return (
-    <React.Fragment>
-      <General />
-      <Font />
-    </React.Fragment>
-  );
-}
-
-export default GlobalStyles;
