+++ src/styled/GlobalStyle/index.js
@@ -1,1 +1,14 @@
-export { default } from './GlobalStyles';
+import React from 'react';
+import General from './General';
+import Font from './Font';
+
+function GlobalStyles() {
+  return (
+    <React.Fragment>
+      <General />
+      <Font />
+    </React.Fragment>
+  );
+}
+
+export default GlobalStyles;
