+++ stories/Styled/Checkbox.stories.js
@@ -1,0 +1,8 @@
+import React from 'react';
+/* eslint-disable import/no-extraneous-dependencies */
+import { storiesOf } from '@storybook/react';
+import Checkbox from '../../src/styled/Checkbox';
+
+storiesOf('Styled|Checkbox', module).add('Checkbox|Simple', () => (
+  <Checkbox name="inputCheckbox" message="Checkbox" />
+));
